move YearWeek to de.metas.util.time
