+++ backend/de.metas.business/src/main/java/de/metas/calendar/plan_optimizer/solver/weekly_capacities/ResourceGroupYearWeek.java
@@ -2,6 +2,7 @@ package de.metas.calendar.plan_optimizer.solver.weekly_capacities;
 
 import com.google.common.base.MoreObjects;
 import de.metas.resource.HumanResourceTestGroupId;
+import de.metas.util.time.YearWeek;
 import lombok.NonNull;
 import lombok.Value;
