+++ backend/de.metas.business/src/main/java/de/metas/calendar/plan_optimizer/solver/weekly_capacities/StepRequiredCapacity.java
@@ -7,6 +7,7 @@ import com.google.common.collect.ImmutableSet;
 import de.metas.calendar.plan_optimizer.domain.Step;
 import de.metas.resource.HumanResourceTestGroupId;
 import de.metas.util.Check;
+import de.metas.util.time.YearWeek;
 import lombok.NonNull;
 import lombok.Value;
 import org.apache.commons.collections4.ListUtils;
